fix: clear the Linux and Windows warning-gate failures reported by CI
Everything the last two CI runs rejected, plus the sites that share each
defect's root cause so the next build does not simply uncover them once the
first failure stops masking its target.

Uninitialised reads. GCC flagged parksmcclellan; MSVC flags the same class of
defect as C4701 elsewhere:
  * ParksMcClellan sets NumBands in four pass-type branches with no else, and
    Remez2 is a direct port of goto-driven Fortran whose flow the compiler
    cannot follow. NUT1, COMP and YNZ are read on paths it cannot prove assign
    them first.
  * FiffEvents read nelem after a tag scan that need not match anything.
  * InvPwlRapMusic read t_val_roh_k, and FiffProducer t_iDiff, before either
    was necessarily assigned.

Shadowing (MSVC C4456/C4458). FiffStream::write_raw_segment declared a
function-scope loop index that every later loop then hid; it is now scoped to
the two loops that need it. FiffInfo::writeToStream shadowed the inherited chs
and nchan members, and FiffEvoked::setInfo's proj parameter hid the member of
the same name it assigns - both renamed rather than left ambiguous.

FiffId::toString used localtime, whose result points into a shared static
buffer: two threads formatting an id concurrently corrupt each other's output.
Switched to the reentrant localtime_r, spelled localtime_s on MSVC, with the
buffer emptied rather than left stale when the conversion fails. This also
clears C4996.

FiffTag::convert_tag_data derived element counts from tag->size(), a size type;
the narrowing to int is safe for any real FIFF tag but has to be explicit,
because size_t is 64-bit on Win64 while int stays 32-bit (C4267).

EventManager::isSharedMemoryInit had an unconditional return after a returning
ifndef branch, making it unreachable whenever IPC is enabled (C4702).

Verified: g++ sweep over all 1410 translation units reports zero diagnostics;
clang build clean; 180/180 ctest pass.

+++ src/libraries/fiff/fiff_tag.cpp
@@ -421,6 +421,9 @@ void FiffTag::convert_matrix_to_file_data(const FiffTag::UPtr& tag)
 //ToDo remove this function by swapping -> define little endian big endian, QByteArray
 void FiffTag::convert_tag_data(const FiffTag::UPtr& tag, int from_endian, int to_endian)
 {
+    // Element counts derived from tag->size(), which is a size type. A FIFF tag
+    // never holds more than INT_MAX elements, so narrowing is safe, but it has
+    // to be explicit: on Win64 size_t is 64-bit while int stays 32-bit.
     int            np;
     int            k,r;//,c;
     char           *offset;
@@ -461,14 +464,14 @@ void FiffTag::convert_tag_data(const FiffTag::UPtr& tag, int from_endian, int to
     case FIFFT_INT :
     case FIFFT_UINT :
     case FIFFT_JULIAN :
-        np = tag->size()/sizeof(fiff_int_t);
+        np = static_cast<int>(tag->size()/sizeof(fiff_int_t));
         for (ithis = (fiff_int_t *)tag->data(), k = 0; k < np; k++, ithis++)
             swap_intp(ithis);
         break;
 
     case FIFFT_LONG :
     case FIFFT_ULONG :
-        np = tag->size()/sizeof(fiff_long_t);
+        np = static_cast<int>(tag->size()/sizeof(fiff_long_t));
         for (lthis = (fiff_long_t *)tag->data(), k = 0; k < np; k++, lthis++)
             swap_longp(lthis);
         break;
@@ -476,21 +479,21 @@ void FiffTag::convert_tag_data(const FiffTag::UPtr& tag, int from_endian, int to
     case FIFFT_SHORT :
     case FIFFT_DAU_PACK16 :
     case FIFFT_USHORT :
-        np = tag->size()/sizeof(fiff_short_t);
+        np = static_cast<int>(tag->size()/sizeof(fiff_short_t));
         for (sthis = (fiff_short_t *)tag->data(), k = 0; k < np; k++, sthis++)
             *sthis = swap_short(*sthis);
         break;
 
     case FIFFT_FLOAT :
     case FIFFT_COMPLEX_FLOAT :
-        np = tag->size()/sizeof(fiff_float_t);
+        np = static_cast<int>(tag->size()/sizeof(fiff_float_t));
         for (fthis = (fiff_float_t *)tag->data(), k = 0; k < np; k++, fthis++)
             swap_floatp(fthis);
         break;
 
     case FIFFT_DOUBLE :
     case FIFFT_COMPLEX_DOUBLE :
-        np = tag->size()/sizeof(fiff_double_t);
+        np = static_cast<int>(tag->size()/sizeof(fiff_double_t));
         for (dthis = (fiff_double_t *)tag->data(), k = 0; k < np; k++, dthis++)
             swap_doublep(dthis);
         break;
